chore: restructure into workspaces
