+++ functions/health/handler.ts
