+++ functions/util/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@frontier/function-util",
+  "version": "1.0.0",
+  "private": true,
+  "main": "sentry.ts"
+}
