+++ packages/db/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@frontier/db",
+  "version": "1.0.0",
+  "private": true,
+  "main": "index.ts"
+}
