+++ packages/testing/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@frontier/testing",
+  "version": "1.0.0",
+  "private": true,
+  "main": "index.ts"
+}
